Changing check for protection to a logic test
This is what performs the check to see if the field should be unprotected or not. Would require a minor detail to a few add-ons, or go the other way around, change the customer notes to "Customer" and check against that.

The idea here is, if the order is tagged by a Customer (ie, during checkout), it's not safe to freely add. Otherwise, if an admin is using admin checkout or another module makes an update (like PayPal  or so on), then it should be safe to allow HTML.